+++ R/comparisons.R
@@ -35,6 +35,7 @@
 #'     * NULL: contrast between TRUE and FALSE
 #'   - Numeric variables:
 #'     * Numeric of length 1: Contrast for a gap of `x`, computed at the observed value plus and minus `x / 2`. For example, estimating a `+1` contrast compares adjusted predictions when the regressor is equal to its observed value minus 0.5 and its observed value plus 0.5.
+#'     * Numeric of length equal to the number of rows in `newdata`: Same as above, but the contrast can be customized for each row of `newdata`.
 #'     * Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the `x` vector.
 #'     * Data frame with the same number of rows as `newdata`, with two columns of "low" and "high" values to compare.
 #'     * Function which accepts a numeric vector and returns a data frame with two columns of "low" and "high" values to compare. See examples below.
@@ -190,6 +191,15 @@
 #'     newdata = "mean",
 #'     hypothesis = lc)
 #'
+#' # Effect of a 1 group-wise standard deviation change
+#' # First we calculate the SD in each group of `cyl`
+#' # Second, we use that SD as the treatment size in the `variables` argument
+#' library(dplyr)
+#' mod <- lm(mpg ~ hp + factor(cyl), mtcars)
+#' tmp <- mtcars %>% 
+#'     group_by(cyl) %>%
+#'     mutate(hp_sd = sd(hp))
+#' avg_comparisons(mod, variables = list(hp = tmp$hp_sd), by = "cyl")
 #'
 #' # `by` argument
 #' mod <- lm(mpg ~ hp * am * vs, data = mtcars)