+++ inst/tinytest/test-variables.R
@@ -80,6 +80,14 @@ b <- avg_comparisons(mod, variables = list(cyl = "revsequential"))
 expect_equal(a$estimate, -1 * b$estimate)
 
 
+# Custom vector
+mod <- lm(mpg ~ hp, mtcars)
+cmp <- avg_comparisons(mod, variables = list(hp = mtcars$cyl), by = "cyl")
+expect_equal(length(unique(cmp$estimate)), 3)
+expect_equal(length(unique(round(cmp$statistic, 8))), 1)
+
+
+
 
 
 rm(list = ls())